pyproject.toml
[project]
name = "comfyui-duffy-nodes"
description = "A custom node pack for ComfyUI built on the V3 Schema."
version = "0.1.0"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.26.0",
    "pillow",
]

[project.urls]
Repository = ""

[tool.comfy]
PublisherId = "duffy"
DisplayName = "Duffy Nodes"
